修复休眠守护服务失控问题，修复远程桌面DPI自适应

+++ SiMay.Core.Standard/Packet/Screen/ScreenBitInfoPacket.cs
@@ -8,9 +8,9 @@ namespace SiMay.Core
 {
     public class ScreenInitBitPacket : EntitySerializerBase
     {
-        public int DpiX { get; set; }
+        public float DpiX { get; set; }
 
-        public int DpiY { get; set; }
+        public float DpiY { get; set; }
         public int Height { get; set; }
         public int Width { get; set; }
